Refactor and rename common part of transport controller
