+++ pkg/transport/cmd/options.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package options
+package cmd
 
 import (
 	"github.com/spf13/pflag"
